+++ Controller/Adminhtml/Form/Save.php
@@ -76,6 +76,7 @@ class Save extends \Alekseon\CustomFormsBuilder\Controller\Adminhtml\Form implem
                 if ($tab) {
                     $tab->addData($tabData);
                 } else {
+                    $tabData['tmp_tab_id'] = $tabId;
                     $form->addFormTab($tabData);
                 }
             }